+++ index.css
@@ -14,3 +14,4 @@
 @import url("./view/settings/index.css");
 @import url("./view/desc/index.css");
 @import url("./view/benchmark/index.css");
+@import url("./view/plot-3d/index.css");
