+++ index.ts
@@ -10,5 +10,4 @@ import './view/code/index.js'
 import './view/checkbox/index.js'
 import './view/settings/index.js'
 import './view/benchmark/index.js'
-
-import './view/3d/index.js'
+import './view/plot-3d/index.js'
