--- tools/color-shapes/const.ts
@@ -1,4 +1,0 @@
-export const L_MAX = 100
-export const C_MAX = 0.37
-export const H_MAX = 360
-export const C_MAX_REC2020 = 0.47
