--- view/3d/mixin.pug
@@ -1,2 +1,0 @@
-mixin plot()
-  canvas.plot-test
