+++ view/layout/index.pug
@@ -1,9 +1,9 @@
 include ../chart/mixin.pug
 include ../field/mixin.pug
 include ../card/mixin.pug
-include ../3d/mixin.pug
+include ../plot-3d/mixin.pug
 
-+plot()
++plot3d()
 
 .layout
   .layout_title
