+++ view/plot-3d/index.css
@@ -1,0 +1,3 @@
+.plot-3d:hover {
+  cursor: grab;
+}
