+++ view/plot-3d/mixin.pug
@@ -1,0 +1,2 @@
+mixin plot3d()
+  canvas.plot-3d
